[v9_10] skip managed-keys reconfig tests in runtime if no crypto [RT #46327]
(cherry picked from commit b093b684a636c0c4b0def51c5ae3c21ec7c06c3a)

+++ bin/tests/system/runtime/tests.sh
@@ -18,41 +18,44 @@ SYSTEMTESTTOP=..
 status=0
 n=0
 
-n=`expr $n + 1`
-echo "I: checking that named refuses to reconfigure if managed-keys-directory is set and not writable ($n)"
-ret=0
-cp -f ns2/named-alt4.conf ns2/named.conf
-$RNDC -c ../common/rndc.conf -s 10.53.0.2 -p 9953 reconfig > rndc.out.$n 2>&1
-grep "failed: permission denied" rndc.out.$n > /dev/null 2>&1 || ret=1
-sleep 1
-grep "managed-keys-directory '.*' is not writable" ns2/named.run > /dev/null 2>&1 || ret=1
-if [ $ret != 0 ]; then echo "I:failed"; fi
-status=`expr $status + $ret`
+if $SHELL ../testcrypto.sh -q
+then
+    n=`expr $n + 1`
+    echo "I: checking that named refuses to reconfigure if managed-keys-directory is set and not writable ($n)"
+    ret=0
+    cp -f ns2/named-alt4.conf ns2/named.conf
+    $RNDC -c ../common/rndc.conf -s 10.53.0.2 -p 9953 reconfig > rndc.out.$n 2>&1
+    grep "failed: permission denied" rndc.out.$n > /dev/null 2>&1 || ret=1
+    sleep 1
+    grep "managed-keys-directory '.*' is not writable" ns2/named.run > /dev/null 2>&1 || ret=1
+    if [ $ret != 0 ]; then echo "I:failed"; fi
+    status=`expr $status + $ret`
 
-n=`expr $n + 1`
-echo "I: checking that named refuses to reconfigure if managed-keys-directory is unset and working directory is not writable ($n)"
-ret=0
-cp -f ns2/named-alt5.conf ns2/named.conf
-$RNDC -c ../common/rndc.conf -s 10.53.0.2 -p 9953 reconfig > rndc.out.$n 2>&1
-grep "failed: permission denied" rndc.out.$n > /dev/null 2>&1 || ret=1
-sleep 1
-grep "working directory '.*' is not writable" ns2/named.run > /dev/null 2>&1 || ret=1
-if [ $ret != 0 ]; then echo "I:failed"; fi
-status=`expr $status + $ret`
+    n=`expr $n + 1`
+    echo "I: checking that named refuses to reconfigure if managed-keys-directory is unset and working directory is not writable ($n)"
+    ret=0
+    cp -f ns2/named-alt5.conf ns2/named.conf
+    $RNDC -c ../common/rndc.conf -s 10.53.0.2 -p 9953 reconfig > rndc.out.$n 2>&1
+    grep "failed:" rndc.out.$n > /dev/null 2>&1 || ret=1
+    sleep 1
+    grep "working directory '.*' is not writable" ns2/named.run > /dev/null 2>&1 || ret=1
+    if [ $ret != 0 ]; then echo "I:failed"; fi
+    status=`expr $status + $ret`
 
-n=`expr $n + 1`
-echo "I: checking that named reconfigures if working directory is not writable but managed-keys-directory is ($n)"
-ret=0
-cp -f ns2/named-alt6.conf ns2/named.conf
-$RNDC -c ../common/rndc.conf -s 10.53.0.2 -p 9953 reconfig > rndc.out.$n 2>&1
-grep "failed: permission denied" rndc.out.$n > /dev/null 2>&1 && ret=1
-if [ $ret != 0 ]; then echo "I:failed"; fi
-status=`expr $status + $ret`
+    n=`expr $n + 1`
+    echo "I: checking that named reconfigures if working directory is not writable but managed-keys-directory is ($n)"
+    ret=0
+    cp -f ns2/named-alt6.conf ns2/named.conf
+    $RNDC -c ../common/rndc.conf -s 10.53.0.2 -p 9953 reconfig > rndc.out.$n 2>&1
+    grep "failed:" rndc.out.$n > /dev/null 2>&1 && ret=1
+    if [ $ret != 0 ]; then echo "I:failed"; fi
+    status=`expr $status + $ret`
 
-echo "I: shutting down existing named"
-[ -s named4.pid ] && kill -15 `cat named4.pid` > /dev/null 2>&1
-[ -s named5.pid ] && kill -15 `cat named5.pid` > /dev/null 2>&1
-[ -s named6.pid ] && kill -15 `cat named6.pid` > /dev/null 2>&1
+    echo "I: shutting down existing named"
+    [ -s named4.pid ] && kill -15 `cat named4.pid` > /dev/null 2>&1
+    [ -s named5.pid ] && kill -15 `cat named5.pid` > /dev/null 2>&1
+    [ -s named6.pid ] && kill -15 `cat named6.pid` > /dev/null 2>&1
+fi
 
 n=`expr $n + 1`
 echo "I: checking that named refuses to start if managed-keys-directory is set and not writable ($n)"
